Fixing issue Prevent user from liking one of their own justifications #16 and updated pipfile.lock file.

+++ evopie/models.py
@@ -377,14 +377,14 @@ class User(UserMixin, DB.Model):
     liked_justifications =DB.relationship('Likes4Justifications', foreign_keys='Likes4Justifications.student_id',
         backref='student', lazy='dynamic')
 
-    def like_justification(self, justification): # TODO #16 Prevent user from liking one of their own justifications
-        if not self.has_liked_justification(justification):
+    def like_justification(self, justification): 
+        if not self.has_liked_justification(justification) and not self.id == justification.student_id:
             like = Likes4Justifications(student_id=self.id, justification_id=justification.id)
             DB.session.add(like)
             DB.session.commit()
 
-    def unlike_justification(self, justification): # TODO #16
-        if self.has_liked_justification(justification):
+    def unlike_justification(self, justification): 
+        if self.has_liked_justification(justification) and not self.id == justification.student_id:
             Likes4Justifications.query.filter_by(
                 student_id=self.id,
                 justification_id=justification.id).delete()
